Move files around, rename files from howtoplay emphasis to article emphasis
Small movement of files here and there, since these article cards and
sections might be re-used in other pagews, its better for them to be
known as article cards/sections.
