+++ src/pages/article.tsx
@@ -1,0 +1,5 @@
+export const getServerSideProps = () => ({ props: { foo: "bar" } });
+
+export default function Page(props: any) {
+  return <p>{JSON.stringify(props)}</p>;
+}
